requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-rest-auth==7.0.0
Django==5.0.9
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-oauth-toolkit==3.0.1
django-stubs==4.2.7
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
graphql-core==3.2.5
idna==3.10
jwcrypto==1.5.6
Markdown==3.7
mypy==1.7.1
mypy-extensions==1.0.0
oauthlib==3.2.2
pillow==11.0.0
promise==2.3
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
strawberry-django-auth==0.377.0
strawberry-graphql==0.247.1
types-pytz==2024.2.0.20241003
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
